+++ src/whatslab/solvers/hand/__init__.py
@@ -1,7 +1,5 @@
 from .controller import HandRetargetController
 from .hand_configs import CONFIG_REGISTRY
-from .kp_retargeter import KPHandRetargeter
-from .retargeter import HandRetargeter
+from .uni_retargeter import UniRetargeter
 
-__all__ = ["HandRetargeter", "KPHandRetargeter", "HandRetargetController",
-           "CONFIG_REGISTRY"]
+__all__ = ["UniRetargeter", "HandRetargetController", "CONFIG_REGISTRY"]